+++ cypress/e2e/granules.cy.js
@@ -1146,5 +1146,46 @@ describe('Dashboard Granules Page', () => {
       cy.get('.search__tip').should('not.exist');
       cy.get('.search__error').should('be.visible');
     });
+
+    it('Should have error and error type columns available to toggle and display correctly', () => {
+      cy.visit('/granules');
+      cy.contains('.table__filters .button__filter', 'Show/Hide Columns').click();
+      cy.get('.table__filters--collapse').should('be.visible');
+
+      cy.contains('.table__filters--filter label', /^Error Type$/).find('span').click();
+      cy.contains('.table__filters--filter label', /^Error$/).find('span').click();
+
+      cy.get('.button__apply-filter').click();
+
+      const noErrorGranuleID = 'test_12345678_123456_metopa_12345_eps_o_coa_1234_ovwcl2';
+      const errorGranuleID = 'MOD09GQ.A2417309.YZ9tCV.006.4640974889044_ca2a8dfe';
+      const unknownErrorGranuleID = 'coastal_12345678_123456_metopa_12345_eps_o_coa_1234_ovwcl3';
+      cy.get(`[data-value="${noErrorGranuleID}"]`).children().as('noErrorColumns');
+      cy.get('@noErrorColumns').eq(4).invoke('text').should('be.eq', '--');
+      cy.get('@noErrorColumns').eq(5).invoke('text').should('be.eq', '--');
+      cy.get(`[data-value="${errorGranuleID}"]`).children().as('errorColumns');
+      cy.get('@errorColumns').eq(4).invoke('text').should('be.eq', 'UnexpectedFileSize');
+      cy.get('@errorColumns').eq(5).invoke('text').should('match', /errorMessage/);
+      cy.get(`[data-value="${unknownErrorGranuleID}"]`).children().as('unknownErrorColumns');
+      cy.get('@unknownErrorColumns').eq(4).invoke('text').should('be.eq', 'Unknown Error');
+      cy.get('@unknownErrorColumns').eq(5).invoke('text').should('match', /None/);
+    });
+
+    it('Should display error if available in granule overview page', () => {
+      const noErrorGranuleID = 'test_12345678_123456_metopa_12345_eps_o_coa_1234_ovwcl2';
+      const errorGranuleID = 'MOD09GQ.A2417309.YZ9tCV.006.4640974889044_ca2a8dfe';
+      const unknownErrorGranuleID = 'coastal_12345678_123456_metopa_12345_eps_o_coa_1234_ovwcl3';
+      cy.visit(`/granules/granule/${errorGranuleID}`);
+      cy.get('.heading--large').should('have.text', `Granule: ${errorGranuleID}`);
+      cy.get('.error__report').should('be.visible');
+      cy.get('.error__report').should('contain.text', 'errorMessage');
+      cy.visit(`/granules/granule/${noErrorGranuleID}`);
+      cy.get('.heading--large').should('have.text', `Granule: ${noErrorGranuleID}`);
+      cy.get('.error__report').should('not.exist');
+      cy.visit(`/granules/granule/${unknownErrorGranuleID}`);
+      cy.get('.heading--large').should('have.text', `Granule: ${unknownErrorGranuleID}`);
+      cy.get('.error__report').should('be.visible');
+      cy.get('.error__report').should('contain.text', 'None');
+    });
   });
 });